Move factory class to its own file
