+++ Unit4/ReportRunnerFactory.cs
@@ -1,0 +1,20 @@
+using System;
+using Unit4.Automation.Interfaces;
+using Unit4.Automation.Model;
+
+namespace Unit4.Automation
+{
+    internal class ReportRunnerFactory
+    {
+        public IRunner Create(IOptions options)
+        {
+            var bcrOptions = options as BcrOptions;
+            if (bcrOptions != null)
+            {
+                return new ReportRunner();
+            }
+
+            return new NullRunner();
+        }
+    }
+}
