+++ Unit4/Unit4.csproj
@@ -102,6 +102,7 @@
     <Compile Include="ReportEngine\Unit4WebConnector.cs" />
     <Compile Include="NullRunner.cs" />
     <Compile Include="ReportRunner.cs" />
+    <Compile Include="ReportRunnerFactory.cs" />
     <Compile Include="BcrReport.cs" />
     <Compile Include="Cache.cs" />
     <Compile Include="Report.cs" />
